rename ScreenerHeader > VueScreenerHeader / ScreenerFooter > VueScreeenerFooter
