+++ src/index.ts
@@ -1,7 +1,7 @@
 import VueScreener from './components/VueScreener.vue'
+import VueScreenerHeader from './components/VueScreenerHeader.vue'
+import VueScreenerFooter from './components/VueScreenerFooter.vue'
 import HeadCell from './components/table/TableHead.vue'
 import DataCell from './components/table/TableData.vue'
-import Header from './components/ScreenerHeader.vue'
-import Footer from './components/ScreenerFooter.vue'
 
-export { VueScreener, HeadCell, DataCell, Header, Footer }
+export { VueScreener, VueScreenerHeader, VueScreenerFooter, HeadCell, DataCell }
