+++ src/stories/1-basic-usage/2-array-of-primitives.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="2. Array of primitives"  source="-">
-    <ScreenerHeader :screener="screener" />
+    <VueScreenerHeader :screener="screener" />
     <VueScreener :screener="screener" />
-    <ScreenerFooter :screener="screener" />
+    <VueScreenerFooter :screener="screener" />
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import { VueScreener, VueScreenerHeader, VueScreenerFooter } from '../../index'
 import primitivesData from '../../fixtures/primitives-data.json'
 import { useScreener } from '../../hooks/use-screener'
