+++ src/stories/2-configure-columns/7-pick-omit-fields.story.vue
@@ -1,13 +1,13 @@
 <template>
   <Story title="7. Pick + omit fields"  source="-">
-    <ScreenerHeader :screener="screener" />
+    <VueScreenerHeader :screener="screener" />
     <VueScreener :screener="screener" />
-    <ScreenerFooter :screener="screener" />
+    <VueScreenerFooter :screener="screener" />
   </Story>
 </template>
 
 <script lang="ts" setup>
-import { VueScreener, Header as ScreenerHeader, Footer as ScreenerFooter } from '../../index'
+import { VueScreener, VueScreenerHeader, VueScreenerFooter } from '../../index'
 import baseData from '../../fixtures/data.json'
 import { useScreener } from '../../hooks/use-screener'
